Relocated home.css and nav.css
